test reward model training with peft, activate test last layer fp16

+++ examples/summarize_rlhf/reward_model/reward_model.py
@@ -14,7 +14,7 @@ class GPTRewardModel(nn.Module):
         # `gpt-neo(x)` models use `hidden_size` attribute names instead of `n_embd``
         self.config.n_embd = self.config.hidden_size if hasattr(self.config, "hidden_size") else self.config.n_embd
         self.transformer = model.transformer
-        self.v_head = nn.Linear(self.config.n_embd, 1, bias=False)
+        self.v_head = nn.Linear(self.config.n_embd, 1, bias=False, dtype=torch.float16)
         self.tokenizer = AutoTokenizer.from_pretrained("EleutherAI/gpt-j-6B")
         self.tokenizer.pad_token = self.tokenizer.eos_token
         self.PAD_ID = self.tokenizer(self.tokenizer.pad_token)["input_ids"][0]
